+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ## Bug-fixes
 
+- [#547 - App Store icon not allowed due to transparency](https://github.com/scriptPilot/app-framework/issues/547)
 - [#562 - Notes in private notes example buggy with two open devices](https://github.com/scriptPilot/app-framework/issues/562)
 - [#569 - File input does not work in native iOS app](https://github.com/scriptPilot/app-framework/issues/569)